+++ src/TGUI/ComboBox.cpp
@@ -222,7 +222,7 @@ namespace tgui
 
         // Make the changes
         m_NrOfItemsToDisplay = nrOfItemsInList;
-        setSize(width, height);
+        setSize(static_cast<float>(width), static_cast<float>(height));
 
         // Check if a scrollbar should be loaded
         if (scrollbarPathname.empty() == false)
@@ -257,7 +257,7 @@ namespace tgui
 
         // Set the height of the combo box
         if (height > m_TopBorder + m_BottomBorder)
-            m_Listbox->setItemHeight(height - m_TopBorder - m_BottomBorder);
+            m_Listbox->setItemHeight(static_cast<unsigned int>(height - m_TopBorder - m_BottomBorder));
         else
             m_Listbox->setItemHeight(10);
 
@@ -838,10 +838,10 @@ namespace tgui
         glGetIntegerv(GL_SCISSOR_BOX, scissor);
 
         // Calculate the clipping area
-        GLint scissorLeft = TGUI_MAXIMUM((globalTranslation.x + m_LeftBorder) * scaleViewX, scissor[0]);
-        GLint scissorTop = TGUI_MAXIMUM((globalTranslation.y + m_TopBorder) * scaleViewY, target.getSize().y - scissor[1] - scissor[3]);
-        GLint scissorRight = TGUI_MINIMUM((globalTranslation.x + m_Listbox->m_Size.x - (m_TextureNormal->getSize().x * curScale.y * (static_cast<float>(m_Listbox->m_ItemHeight) / m_TextureNormal->getSize().y)) - m_RightBorder) * scaleViewX, scissor[0] + scissor[2]);
-        GLint scissorBottom = TGUI_MINIMUM((globalTranslation.y + m_Listbox->m_Size.y - m_BottomBorder) * scaleViewY, target.getSize().y - scissor[1]);
+        GLint scissorLeft = TGUI_MAXIMUM(static_cast<GLint>((globalTranslation.x + m_LeftBorder) * scaleViewX), scissor[0]);
+        GLint scissorTop = TGUI_MAXIMUM(static_cast<GLint>((globalTranslation.y + m_TopBorder) * scaleViewY), static_cast<GLint>(target.getSize().y) - scissor[1] - scissor[3]);
+        GLint scissorRight = TGUI_MINIMUM(static_cast<GLint>((globalTranslation.x + m_Listbox->m_Size.x - (m_TextureNormal->getSize().x * curScale.y * (static_cast<float>(m_Listbox->m_ItemHeight) / m_TextureNormal->getSize().y)) - m_RightBorder) * scaleViewX), scissor[0] + scissor[2]);
+        GLint scissorBottom = TGUI_MINIMUM(static_cast<GLint>((globalTranslation.y + m_Listbox->m_Size.y - m_BottomBorder) * scaleViewY), static_cast<GLint>(target.getSize().y) - scissor[1]);
 
         // If the object outside the window then don't draw anything
         if (scissorRight < scissorLeft)
@@ -887,7 +887,7 @@ namespace tgui
         if (m_ShowList)
         {
             // Set the listbox to the correct position and draw it
-            states.transform = oldTransform.translate(0, m_Listbox->m_ItemHeight + m_TopBorder + m_BottomBorder);
+            states.transform = oldTransform.translate(0, static_cast<float>(m_Listbox->m_ItemHeight + m_TopBorder + m_BottomBorder));
             target.draw(*m_Listbox, states);
         }
     }
